harvester-api/requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
Flask==3.0.2
idna==3.6
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
